@@ -1,10 +1,8 @@
-import 'dart:io';
 import 'package:blue_thermal_printer_example/testprint.dart';
 import 'package:flutter/material.dart';
 import 'dart:async';
 import 'package:blue_thermal_printer/blue_thermal_printer.dart';
 import 'package:flutter/services.dart';
-import 'package:path_provider/path_provider.dart';
 
 void main() => runApp(new MyApp());
 
@@ -17,39 +15,22 @@ class _MyAppState extends State<MyApp> {
   BlueThermalPrinter bluetooth = BlueThermalPrinter.instance;
 
   List<BluetoothDevice> _devices = [];
-  BluetoothDevice _device;
+  BluetoothDevice? _device;
   bool _connected = false;
-  String pathImage;
-  TestPrint testPrint;
+  TestPrint testPrint = TestPrint();
 
   @override
   void initState() {
     super.initState();
     initPlatformState();
-    initSavetoPath();
-    testPrint = TestPrint();
-  }
-
-  initSavetoPath() async {
-    //read and write
-    //image max 300px X 300px
-    final filename = 'yourlogo.png';
-    var bytes = await rootBundle.load("assets/images/yourlogo.png");
-    String dir = (await getApplicationDocumentsDirectory()).path;
-    writeToFile(bytes, '$dir/$filename');
-    setState(() {
-      pathImage = '$dir/$filename';
-    });
   }
 
   Future<void> initPlatformState() async {
-    bool isConnected = await bluetooth.isConnected;
+    bool? isConnected = await bluetooth.isConnected;
     List<BluetoothDevice> devices = [];
     try {
       devices = await bluetooth.getBondedDevices();
-    } on PlatformException {
-      // TODO - Error
-    }
+    } on PlatformException {}
 
     bluetooth.onStateChanged().listen((state) {
       switch (state) {
@@ -112,7 +93,7 @@ class _MyAppState extends State<MyApp> {
       _devices = devices;
     });
 
-    if (isConnected) {
+    if (isConnected == true) {
       setState(() {
         _connected = true;
       });
@@ -150,7 +131,8 @@ class _MyAppState extends State<MyApp> {
                     Expanded(
                       child: DropdownButton(
                         items: _getDeviceItems(),
-                        onChanged: (value) => setState(() => _device = value),
+                        onChanged: (BluetoothDevice? value) =>
+                            setState(() => _device = value),
                         value: _device,
                       ),
                     ),
@@ -193,7 +175,7 @@ class _MyAppState extends State<MyApp> {
                   child: ElevatedButton(
                     style: ElevatedButton.styleFrom(primary: Colors.brown),
                     onPressed: () {
-                      testPrint.sample(pathImage);
+                      testPrint.sample();
                     },
                     child: Text('PRINT TEST',
                         style: TextStyle(color: Colors.white)),
@@ -216,7 +198,7 @@ class _MyAppState extends State<MyApp> {
     } else {
       _devices.forEach((device) {
         items.add(DropdownMenuItem(
-          child: Text(device.name),
+          child: Text(device.name ?? ""),
           value: device,
         ));
       });
@@ -225,30 +207,23 @@ class _MyAppState extends State<MyApp> {
   }
 
   void _connect() {
-    if (_device == null) {
-      show('No device selected.');
-    } else {
+    if (_device != null) {
       bluetooth.isConnected.then((isConnected) {
-        if (!isConnected) {
-          bluetooth.connect(_device).catchError((error) {
+        if (isConnected == true) {
+          bluetooth.connect(_device!).catchError((error) {
             setState(() => _connected = false);
           });
           setState(() => _connected = true);
         }
       });
+    } else {
+      show('No device selected.');
     }
   }
 
   void _disconnect() {
     bluetooth.disconnect();
     setState(() => _connected = false);
-  }
-
-//write to app path
-  Future<void> writeToFile(ByteData data, String path) {
-    final buffer = data.buffer;
-    return new File(path).writeAsBytes(
-        buffer.asUint8List(data.offsetInBytes, data.lengthInBytes));
   }
 
   Future show(
